+++ src/components/Toaster/index.stories.js
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { v4 } from 'uuid';
 
 import { Toaster } from 'components/Toaster';
-import Toast, { TYPES } from './Toast';
+import Toast from './Toast';
 
 export const Primary = (args) => {
   const [isToastShowing, setIsToastShowing] = useState(false);
@@ -14,9 +14,9 @@ export const Primary = (args) => {
       <Toaster
         {...args}
         isOpen={isToastShowing}
-        onDismiss={() => setIsToastShowing(false)}
-        toastProps={{
+        toastInstance={{
           content: 'hey',
+          onDismiss: () => setIsToastShowing(false),
           type: 'info',
           progress: 42,
           dismissOnAction: true,
@@ -27,34 +27,42 @@ export const Primary = (args) => {
   );
 };
 
-export const rendering = () => {
+export const playground = () => {
   const defaults = { id: v4(), isOpen: false, isAutoDismiss: true };
   const [params, setParams] = useState(defaults);
-  const [isShadow, setIsShadow] = useState(false);
   const dismissToast = () => {
     setParams({ ...params, isOpen: false });
-  };
-  const randomToastType = () => {
-    const randomNumber = Math.floor(Math.random() * TYPES.length);
-    return TYPES[randomNumber];
   };
   const randomNum = () => Math.floor(Math.random() * 100);
 
   const { id, type, isOpen, content, isAutoDismiss, actionLabel, progress } = params;
   return (
-    <div className={isShadow ? 'elevation--1 padding--all' : 'padding--all'}>
-      <button onClick={() => setIsShadow(!isShadow)}>
-        {isShadow ? 'Hide Arbitrary Style' : 'Show Arbitrary Style'}
-      </button>
+    <div className="padding--all">
+      <h4>Core:</h4>
       <button onClick={() => setParams({ ...params, isOpen: true })}>Open Toast</button>
       <button onClick={() => setParams({ ...params, isOpen: false })}>Close Toast</button>
       <button onClick={() => setParams({ ...params, isAutoDismiss: !isAutoDismiss })}>
         Toggle isAutoDismiss {isAutoDismiss ? 'off' : 'on'}
       </button>
       <button onClick={() => setParams({ ...params, id: v4() })}>New ID</button>
-      <button onClick={() => setParams({ ...params, type: randomToastType() })}>
-        Random Toast Type
+      <button onClick={() => setParams({ ...defaults })}>Reset all</button>
+      <br />
+      <br />
+      <h4>Types:</h4>
+      <button onClick={() => setParams({ ...params, type: 'info' })}>info</button>
+      <button onClick={() => setParams({ ...params, type: 'warn' })}>warn</button>
+      <button onClick={() => setParams({ ...params, type: 'error' })}>error</button>
+      <button onClick={() => setParams({ ...params, type: 'success' })}>success</button>
+      <button onClick={() => setParams({ ...params, type: 'progress' })}>progress</button>
+      <button onClick={() => setParams({ ...params, progress: randomNum() })}>
+        Add random progress
       </button>
+      <button onClick={() => setParams({ ...params, progress: null })}>
+        Remove progress
+      </button>
+      <br />
+      <br />
+      <h4>Modify Content:</h4>
       <button
         onClick={() =>
           setParams({
@@ -65,16 +73,21 @@ export const rendering = () => {
       >
         Change Action Label
       </button>
-      <button onClick={() => setParams({ ...params, content: <b>hey</b> })}>
+      <button onClick={() => setParams({ ...params, content: 'hello world' })}>
+        Add content
+      </button>
+      <button onClick={() => setParams({ ...params, content: <b>hello world</b> })}>
         Add HTML content
+      </button>
+      <button onClick={() => setParams({ ...params, content: null })}>
+        Remove content
       </button>
       <button onClick={() => setParams({ ...params, actionLabel: null })}>
         Remove Action Label
       </button>
-      <button onClick={() => setParams({ ...params, progress: randomNum() })}>
-        Add random progress
-      </button>
       <br />
+      <br />
+      <h4>Example usages:</h4>
       <button
         onClick={() =>
           setParams({
@@ -86,7 +99,7 @@ export const rendering = () => {
           })
         }
       >
-        Error Toast (new id)
+        Error Toast (animate)
       </button>
       <button
         onClick={() =>
@@ -99,7 +112,7 @@ export const rendering = () => {
           })
         }
       >
-        Info Toast (new id)
+        Info Toast (animate)
       </button>
       <button
         onClick={() =>
@@ -112,9 +125,8 @@ export const rendering = () => {
           })
         }
       >
-        Error Toast (no new id)
+        Error Toast (do not animate)
       </button>
-      <button onClick={() => setParams({ ...defaults })}>Reset all</button>
       <br />
       <br />
       <b>NOTE: this is state is EXTERNAL to toaster and is being passed in</b>
@@ -122,9 +134,9 @@ export const rendering = () => {
       <Toaster
         id={id}
         isOpen={isOpen}
-        onDismiss={dismissToast}
         isAutoDismiss={isAutoDismiss}
-        toastProps={{
+        toastInstance={{
+          onDismiss: dismissToast,
           actionLabel,
           progress,
           type,
@@ -133,6 +145,15 @@ export const rendering = () => {
       />
     </div>
   );
+};
+
+playground.parameters = {
+  docs: {
+    description: {
+      story:
+        'This is a playground where you can enable certain functionality and see how Toaster reacts in real time, while being able to visualize the exact prop structure to generate it.',
+    },
+  },
 };
 
 export const variations = () => (
@@ -200,6 +221,17 @@ export const variations = () => (
     />
   </React.Fragment>
 );
+
+variations.parameters = {
+  docs: {
+    description: {
+      story: 'This is merely to give you a better visual idea of all the variations.',
+    },
+    source: {
+      code: 'Hidden. This is merely to give you a visual idea of the themes available.',
+    },
+  },
+};
 
 export default {
   component: Toaster,